Dss set schema uses api (#74)
* Fix dss set_schema, it must use API not Dataset object

* Correct type mapping

bigint is LongType, int is IntegerType, in spark.

+++ tests/dataframesource/sources/test_s3source.py
@@ -4,7 +4,7 @@ import pytest
 from birgitta import spark
 from birgitta.dataframe import dataframe, dfdiff
 from birgitta.dataframesource.sources.s3source import S3Source  # noqa F401
-from pyspark.sql.types import IntegerType
+from pyspark.sql.types import LongType
 from pyspark.sql.types import StringType
 from pyspark.sql.types import StructField
 from pyspark.sql.types import StructType
@@ -17,7 +17,7 @@ def spark_session():
 
 fixtures_schema = StructType([
     StructField('letter', StringType()),
-    StructField('number', IntegerType())
+    StructField('number', LongType())
 ])
