+++ tests/schema/spark/test_print_df_rows.py
@@ -1,6 +1,6 @@
 import pytest
 from birgitta.schema.spark import print_df_rows
-from pyspark.sql.types import IntegerType
+from pyspark.sql.types import LongType
 from pyspark.sql.types import StringType
 from pyspark.sql.types import StructField
 from pyspark.sql.types import StructType
@@ -8,7 +8,7 @@ from pyspark.sql.types import StructType
 
 fixtures_schema = StructType([
     StructField('letter', StringType()),
-    StructField('number', IntegerType())
+    StructField('number', LongType())
 ])
